First approach to generalize diagnostic info population
