+++ function-aws-api-proxy/src/main/java/io/micronaut/function/aws/proxy/MicronautLambdaHandler.java
@@ -65,6 +65,7 @@ public class MicronautLambdaHandler implements RequestHandler<AwsProxyRequest, A
 
     @Override
     public AwsProxyResponse handleRequest(AwsProxyRequest input, Context context) {
+
         return handler.proxy(input, context);
     }
