+++ function-aws/src/main/java/io/micronaut/function/aws/MicronautRequestHandler.java
@@ -15,19 +15,21 @@
  */
 package io.micronaut.function.aws;
 
-import com.amazonaws.services.lambda.runtime.*;
-import io.micronaut.core.annotation.NonNull;
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestHandler;
 import io.micronaut.context.ApplicationContext;
 import io.micronaut.context.ApplicationContextBuilder;
+import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.convert.ArgumentConversionContext;
 import io.micronaut.core.convert.ConversionContext;
 import io.micronaut.core.convert.ConversionError;
 import io.micronaut.core.reflect.GenericTypeUtils;
 import io.micronaut.core.util.ArrayUtils;
-import io.micronaut.core.util.StringUtils;
 import io.micronaut.function.executor.AbstractFunctionExecutor;
-import org.slf4j.MDC;
+
 import java.util.Optional;
+
+import static io.micronaut.function.aws.DiagnosticInfoPopulator.registerContextBeans;
 
 /**
  * <p>An Amazon Lambda {@link RequestHandler} implementation for Micronaut {@link io.micronaut.function.FunctionBean}</p>.
@@ -39,21 +41,9 @@ import java.util.Optional;
  */
 public abstract class MicronautRequestHandler<I, O> extends AbstractFunctionExecutor<I, O, Context> implements RequestHandler<I, O>, MicronautLambdaContext {
 
-    public static final String ENV_X_AMZN_TRACE_ID = "_X_AMZN_TRACE_ID";
-
-    // See: https://github.com/aws/aws-xray-sdk-java/issues/251
-    public static final String LAMBDA_TRACE_HEADER_PROP = "com.amazonaws.xray.traceHeader";
-
-    public static final String MDC_DEFAULT_AWS_REQUEST_ID = "AWSRequestId";
-    public static final String MDC_DEFAULT_FUNCTION_NAME = "AWSFunctionName";
-    public static final String MDC_DEFAULT_FUNCTION_VERSION = "AWSFunctionVersion";
-    public static final String MDC_DEFAULT_FUNCTION_ARN = "AWSFunctionArn";
-    public static final String MDC_DEFAULT_FUNCTION_MEMORY_SIZE = "AWSFunctionMemoryLimit";
-    public static final String MDC_DEFAULT_FUNCTION_REMAINING_TIME = "AWSFunctionRemainingTime";
-    public static final String MDC_DEFAULT_XRAY_TRACE_ID = "AWS-XRAY-TRACE-ID";
-
     @SuppressWarnings("unchecked")
     private final Class<I> inputType = initTypeArgument();
+    protected final DiagnosticInfoPopulator diagnosticInfoPopulator;
 
     /**
      * Default constructor; will initialize a suitable {@link ApplicationContext} for
@@ -62,6 +52,7 @@ public abstract class MicronautRequestHandler<I, O> extends AbstractFunctionExec
     public MicronautRequestHandler() {
         buildApplicationContext(null);
         injectIntoApplicationContext();
+        this.diagnosticInfoPopulator = applicationContext.getBean(DiagnosticInfoPopulator.class);
     }
 
     /**
@@ -71,81 +62,27 @@ public abstract class MicronautRequestHandler<I, O> extends AbstractFunctionExec
     public MicronautRequestHandler(ApplicationContext applicationContext) {
         this.applicationContext = applicationContext;
         injectIntoApplicationContext();
+        this.diagnosticInfoPopulator = applicationContext.getBean(DiagnosticInfoPopulator.class);
     }
 
     private void injectIntoApplicationContext() {
         applicationContext.inject(this);
+        applicationContext.inject(new DiagnosticInfoPopulator());
     }
 
     @Override
     public final O handleRequest(I input, Context context) {
         if (context != null) {
             registerContextBeans(context, applicationContext);
-            populateMappingDiagnosticContextValues(context);
+            diagnosticInfoPopulator.populateMappingDiagnosticContextValues(context);
         }
-        populateMappingDiagnosticContextWithXrayTraceId();
+        diagnosticInfoPopulator.populateMappingDiagnosticContextWithXrayTraceId();
         if (!inputType.isInstance(input)) {
             input = convertInput(input);
         }
         return this.execute(input);
     }
 
-    /**
-     * @see <a href="https://docs.aws.amazon.com/lambda/latest/dg/java-logging.html">AWS Lambda function logging in Java</a>
-     * @param context The Lambda execution environment context object.
-     */
-    protected void populateMappingDiagnosticContextValues(@NonNull Context context) {
-        if (context.getAwsRequestId() != null) {
-            mdcput(MDC_DEFAULT_AWS_REQUEST_ID, context.getAwsRequestId());
-        }
-        if (context.getFunctionName() != null) {
-            mdcput(MDC_DEFAULT_FUNCTION_NAME, context.getFunctionName());
-        }
-        if (context.getFunctionVersion() != null) {
-            mdcput(MDC_DEFAULT_FUNCTION_VERSION, context.getFunctionVersion());
-        }
-        if (context.getInvokedFunctionArn() != null) {
-            mdcput(MDC_DEFAULT_FUNCTION_ARN, context.getInvokedFunctionArn());
-        }
-        mdcput(MDC_DEFAULT_FUNCTION_MEMORY_SIZE, String.valueOf(context.getMemoryLimitInMB()));
-        mdcput(MDC_DEFAULT_FUNCTION_REMAINING_TIME, String.valueOf(context.getRemainingTimeInMillis()));
-    }
-
-    /**
-     * Put a diagnostic context value.
-     * @param key non-null key
-     * @param val value to put in the map
-     * @throws IllegalArgumentException in case the "key" parameter is null
-     */
-    protected void mdcput(@NonNull String key, @NonNull String val) throws IllegalArgumentException {
-        MDC.put(key, val);
-    }
-
-    /**
-     * Populate MDC with XRay Trace ID if is able to parse it.
-     */
-    protected void populateMappingDiagnosticContextWithXrayTraceId() {
-        parseXrayTraceId().ifPresent(xrayTraceId -> mdcput(MDC_DEFAULT_XRAY_TRACE_ID, xrayTraceId));
-    }
-
-    /**
-     * Parses XRay Trace ID from _X_AMZN_TRACE_ID environment variable.
-     * @see <a href="https://docs.aws.amazon.com/xray/latest/devguide/xray-sdk-java-configuration.html">Trace ID injection into logs</a>
-     * @return Trace id or empty if not found
-     */
-    @NonNull
-    protected static Optional<String> parseXrayTraceId() {
-        String lambdaTraceHeaderKey = System.getenv(ENV_X_AMZN_TRACE_ID);
-        lambdaTraceHeaderKey = StringUtils.isNotEmpty(lambdaTraceHeaderKey) ? lambdaTraceHeaderKey
-                : System.getProperty(LAMBDA_TRACE_HEADER_PROP);
-        if (lambdaTraceHeaderKey != null) {
-            String[] arr = lambdaTraceHeaderKey.split(";");
-            if (arr.length >= 1) {
-                return Optional.of(arr[0].replace("Root=", ""));
-            }
-        }
-        return Optional.empty();
-    }
 
     /**
      * Converts the input the required type. Subclasses can override to provide custom conversion.
@@ -177,28 +114,6 @@ public abstract class MicronautRequestHandler<I, O> extends AbstractFunctionExec
     @NonNull
     protected ApplicationContextBuilder newApplicationContextBuilder() {
         return new LambdaApplicationContextBuilder();
-    }
-
-    /**
-     * Register the beans in the application.
-     *
-     * @param context context
-     * @param applicationContext application context
-     */
-    static void registerContextBeans(Context context, ApplicationContext applicationContext) {
-        applicationContext.registerSingleton(context);
-        LambdaLogger logger = context.getLogger();
-        if (logger != null) {
-            applicationContext.registerSingleton(logger);
-        }
-        ClientContext clientContext = context.getClientContext();
-        if (clientContext != null) {
-            applicationContext.registerSingleton(clientContext);
-        }
-        CognitoIdentity identity = context.getIdentity();
-        if (identity != null) {
-            applicationContext.registerSingleton(identity);
-        }
     }
 
     private Class initTypeArgument() {
