+++ function-aws/src/main/java/io/micronaut/function/aws/MicronautRequestStreamHandler.java
@@ -27,7 +27,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import static io.micronaut.function.aws.MicronautRequestHandler.registerContextBeans;
+import static io.micronaut.function.aws.DiagnosticInfoPopulator.registerContextBeans;
 
 /**
  * <p>An implementation of the {@link RequestStreamHandler} for Micronaut</p>.
